refactor: Move client to client.rs
