+++ influxdb2_client/src/api/mod.rs
@@ -3,3 +3,4 @@ pub mod buckets;
 pub mod query;
 pub mod ready;
 pub mod setup;
+pub mod write;
