+++ influxdb2_client/src/api/ready.rs
@@ -2,10 +2,9 @@
 //!
 //! Check readiness of an InfluxDB instance at startup
 
+use crate::{Client, Http, RequestError, ReqwestProcessing};
 use reqwest::{Method, StatusCode};
 use snafu::ResultExt;
-
-use crate::{Client, Http, RequestError, ReqwestProcessing};
 
 impl Client {
     /// Get the readiness of an instance at startup
